add SearchBar for Groups
