--- VK_app/AvailGroupList/AvailGroupCell.swift
@@ -1,21 +1,0 @@
-//
-//  AvailGroupCell.swift
-//  VK_app
-//
-//  Created by Nikolay.Trofimov on 28/12/2018.
-//  Copyright © 2018 Nikolay.Trofimov. All rights reserved.
-//
-
-import UIKit
-
-class AvailGroupCell: UITableViewCell {
-
-  @IBOutlet weak var userGroupImage: UIImageView!
-  @IBOutlet weak var userGroupLabel: UILabel!
-  
-
-  public func configure(with groupList: Group) {
-    // TODO: создать фунцию заполения данных
-  }
-  
-}
